backend/requirements/base.txt
asgiref==3.8.1
Django==5.2.3
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
